refactor(sales): split SalesOrders (654→144) em hook + componentes
Extrai a lógica (2 infinite queries sales+afiação, merge, profiles, infinite
scroll, deleteOrder/deleteSelected optimistic com rollback, filteredOrders) para
o hook useSalesOrders e quebra o JSX em SalesOrdersToolbar e SalesOrderCard.
Tipos/helpers (decodeHtml, statusLabels, PAGE_SIZE) em types.ts. Comportamento
1:1 — incluindo o optimistic delete/rollback (referência viva do CLAUDE.md). +12 testes.
